Separate chunk benchmarks
