+++ benchmarks/src/main/scala/zio/stream/StreamExperiments.scala
@@ -42,35 +42,4 @@ class StreamExperiments {
 
     unsafeRun(result)
   }
-
-  @Benchmark
-  def zioChunkFilterMapSum = {
-    val stream = ZStream
-      .fromChunk(Chunk(1 to 100: _*))
-      .forever
-      .take(count)
-      .filter(_ % 2 == 0)
-      .map(_.toLong)
-
-    val sink   = ZSink.foldLeftChunks(0L)((s, as: Chunk[Long]) => as.fold(s)(_ + _))
-    val result = stream.run(sink)
-
-    unsafeRun(result)
-  }
-
-  @Benchmark
-  def newChunkFilterMapSum = {
-    val stream = XStream
-      .apply(Chunk(1 to 100: _*))
-      .forever
-      .pipe(XTransducer.take[Int](count).chunked)
-      .pipe(XTransducer.filter[Int](_ % 2 == 0))
-      .pipe(XTransducer.map[Int, Long](_.toLong).chunked)
-
-    val sink   = XSink.sum[Long].chunked
-    val result = stream.run(sink)
-
-    unsafeRun(result)
-  }
-
 }
